backend/requirements.txt
cmake==3.28.1               # Thêm phiên bản cụ thể để tăng tính ổn định
fastapi==0.115.0           
orjson==3.10.0             
pandas==2.2.0               
pydantic==2.9.0   
boto3==1.37.9          
python-dotenv==1.0.1 
requests==2.32.3            
uvicorn==0.30.0             
faiss-cpu==1.8.0            
stripe==10.8.0 
openai==1.10.0              # Giữ nguyên phiên bản
anthropic==0.34.2          
httpx==0.26.0  
web3==6.20.1           
apscheduler==3.11.0